security transactions

+++ src/main/java/com/gni/banking/Repository/UserRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 public interface UserRepository extends CrudRepository<User,Long> {
 
 	@Query("SELECT dayLimit FROM User WHERE id = :userId")
-	double getDayLimitById(@Param("userId") int userId);
+	double getDayLimitById(@Param("userId") long userId);
 
 	List<User> findByFirstNameAndLastName(String firstName, String lastName);
     // Add custom repository methods if needed
